+++ eth/tracers/internal/tracetest/testdata/call_tracer/inner_throw_outer_revert.json
