+++ eth/tracers/internal/tracetest/testdata/call_tracer/selfdestruct.json
