+++ eth/tracers/internal/tracetest/testdata/call_tracer/throw.json
