+++ eth/tracers/internal/tracetest/testdata/call_tracer_legacy/inner_instafail.json
