+++ eth/tracers/internal/tracetest/testdata/call_tracer_legacy/inner_throw_outer_revert.json
