+++ eth/tracers/internal/tracetest/testdata/call_tracer_legacy/selfdestruct.json
