+++ eth/tracers/js/internal/tracers/assets.go
