+++ eth/tracers/js/internal/tracers/bigram_tracer.js
