+++ eth/tracers/js/internal/tracers/opcount_tracer.js
